+++ ci/install-packages.sh
@@ -31,6 +31,7 @@ else
         apt install ./libgraphviz4_${GV_VERSION}-1_amd64.deb
         apt install ./libgraphviz-dev_${GV_VERSION}-1_amd64.deb
         apt install ./graphviz_${GV_VERSION}-1_amd64.deb
+        apt install ./libgv-tcl_${GV_VERSION}-1_amd64.deb
     elif [ "${ID}" = "Darwin" ]; then
         tar xf ${DIR}/graphviz-${GV_VERSION}-${ARCH}.tar.gz
         sudo cp -rp build/* /usr/local