+++ tclpkg/tclpathplan/tclpathplan.c
@@ -158,6 +158,9 @@ static void expandPercentsEval(Tcl_Interp *interp, char *before,
 	    agxbputc(&scripts, before[1]);
 	    break;
 	}
+	if (before[1] == '\0') {
+	    break;
+	}
 	before += 2;
     }
     const char *script_value = agxbuse(&scripts);
@@ -755,9 +758,9 @@ vgpanecmd(ClientData clientData, Tcl_Interp * interp, int argc,
 	return TCL_ERROR;
 
     } else if (strcmp(argv[1], "triangulate") == 0) {
-	if (argc < 2) {
+	if (argc < 3) {
 	    Tcl_AppendResult(interp, "wrong # args: should be \"", argv[0],
-			     " id ", NULL);
+			     " ", argv[1], " id\"", NULL);
 	    return TCL_ERROR;
 	}
 
@@ -768,7 +771,13 @@ vgpanecmd(ClientData clientData, Tcl_Interp * interp, int argc,
 
 	for (size_t i = 0; i < polys_size(&vgp->poly); i++) {
 	    if (polys_get(&vgp->poly, i).id == polyid) {
-		Ptriangulate(&polys_at(&vgp->poly, i)->boundary, triangle_callback, vgp);
+		Ppoly_t *polygon = &polys_at(&vgp->poly, i)->boundary;
+		if (polygon->pn < 3) {
+		    Tcl_AppendResult(interp, "polygon ", argv[2], " has fewer than 3 points "
+		                     "and thus cannot be triangulated", NULL);
+		    return TCL_ERROR;
+		}
+		Ptriangulate(polygon, triangle_callback, vgp);
 		return TCL_OK;
 	    }
 	}
@@ -866,6 +875,7 @@ vgpane(ClientData clientData, Tcl_Interp * interp, int argc, const char *argv[])
     vgpanes_append(&vgpaneTable, vg);
     vgpane_t *const vgp = vgpanes_back(&vgpaneTable);
     vgp->index = vgpanes_size(&vgpaneTable) - 1;
+    vgp->valid = true;
 
     agxbuf buffer = {0};
     agxbprint(&buffer, HANDLE_FORMAT, vgp->index);